renaming instrument_view and changing value_type to dict
