+++ nexus_constructor/instrument_view/axis_colors.py
