+++ nexus_constructor/instrument_view/instrument_view.py
@@ -8,15 +8,18 @@ from PySide2.QtCore import QRectF
 from PySide2.QtGui import QVector3D, QColor
 from PySide2.QtWidgets import QWidget, QVBoxLayout
 
-from nexus_constructor.spacialview.gnomon import Gnomon
-from nexus_constructor.spacialview.instrument_view_axes import InstrumentViewAxes
-from nexus_constructor.spacialview.instrument_zooming_3d_window import (
+from nexus_constructor.instrument_view.gnomon import Gnomon
+from nexus_constructor.instrument_view.instrument_view_axes import InstrumentViewAxes
+from nexus_constructor.instrument_view.instrument_zooming_3d_window import (
     InstrumentZooming3DWindow,
 )
 from nexus_constructor.model.geometry import OFFGeometry
 from nexus_constructor.model.instrument import SAMPLE_NAME
-from nexus_constructor.spacialview.off_renderer import OffMesh
-from nexus_constructor.spacialview.qentity_utils import create_qentity, create_material
+from nexus_constructor.instrument_view.off_renderer import OffMesh
+from nexus_constructor.instrument_view.qentity_utils import (
+    create_qentity,
+    create_material,
+)
 
 
 class InstrumentView(QWidget):
