+++ nexus_constructor/instrument_view/instrument_zooming_3d_window.py
