+++ nexus_constructor/instrument_view/neutron_animation_controller.py
