+++ tests/test_neutron_animation_controller.py
@@ -1,5 +1,5 @@
 from PySide2.QtGui import QMatrix4x4
-from nexus_constructor.spacialview.neutron_animation_controller import (
+from nexus_constructor.instrument_view.neutron_animation_controller import (
     NeutronAnimationController,
 )
 from mock import Mock
