+++ tests/test_qentity_utils.py
@@ -1,6 +1,9 @@
 from mock import Mock, call, patch
 
-from nexus_constructor.spacialview.qentity_utils import create_material, create_qentity
+from nexus_constructor.instrument_view.qentity_utils import (
+    create_material,
+    create_qentity,
+)
 
 
 @patch("PySide2.Qt3DExtras.Qt3DExtras.QPhongMaterial", return_value=Mock())
